+++ example-crates/origin-start-no-alloc/src/main.rs
@@ -2,23 +2,9 @@
 
 #![no_std]
 #![no_main]
-#![allow(internal_features)]
-#![feature(lang_items)]
-#![feature(core_intrinsics)]
 
-extern crate compiler_builtins;
-
-use atomic_dbg::{dbg, eprintln};
+use atomic_dbg::eprintln;
 use origin::program;
-
-#[panic_handler]
-fn panic(panic: &core::panic::PanicInfo<'_>) -> ! {
-    dbg!(panic);
-    core::intrinsics::abort();
-}
-
-#[lang = "eh_personality"]
-extern "C" fn eh_personality() {}
 
 #[no_mangle]
 unsafe fn origin_main(_argc: usize, _argv: *mut *mut u8, _envp: *mut *mut u8) -> i32 {