+++ src/arch/x86_64.rs
@@ -1,10 +1,12 @@
 //! Architecture-specific assembly code.
 
+#[allow(unused_imports)]
 use core::arch::asm;
 #[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
 #[cfg(relocation_model = "pic")]
 use linux_raw_sys::elf::{Elf_Dyn, Elf_Ehdr};
 #[cfg(feature = "signal")]
+#[cfg(test)]
 use linux_raw_sys::general::__NR_rt_sigreturn;
 #[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
 #[cfg(relocation_model = "pic")]
@@ -16,35 +18,33 @@ use {
     rustix::thread::RawPid,
 };
 
-/// The program entry point.
-///
-/// # Safety
-///
-/// This function must never be called explicitly. It is the first thing
-/// executed in the program, and it assumes that memory is laid out according
-/// to the operating system convention for starting a new program.
 #[cfg(feature = "origin-start")]
-#[naked]
-#[no_mangle]
-pub(super) unsafe extern "C" fn _start() -> ! {
+naked!(
+    "
+    The program entry point.
+
+    # Safety
+
+    This function must never be called explicitly. It is the first thing
+    executed in the program, and it assumes that memory is laid out according
+    to the operating system convention for starting a new program.
+    ";
+    pub(super) fn _start() -> !;
+
     // Jump to `entry`, passing it the initial stack pointer value as an
     // argument, a null return address, a null frame pointer, and an aligned
     // stack pointer. On many architectures, the incoming frame pointer is
     // already null.
-    asm!(
-        "mov rdi, rsp", // Pass the incoming `rsp` as the arg to `entry`.
-        "push rbp",     // Set the return address to zero.
-        "jmp {entry}",  // Jump to `entry`.
-        entry = sym super::program::entry,
-        options(noreturn),
-    )
-}
+    "mov rdi, rsp", // Pass the incoming `rsp` as the arg to `entry`.
+    "push rbp",     // Set the return address to zero.
+    "jmp {entry}";  // Jump to `entry`.
+    entry = sym super::program::entry;
+    options(noreturn)
+);
 
 /// Abort the process without involving any panic handling code.
 ///
 /// This is a stable equivalent to `core::intrinsics::abort()`.
-#[cfg(all(feature = "experimental-relocate", feature = "origin-start"))]
-#[cfg(relocation_model = "pic")]
 pub(super) fn abort() -> ! {
     unsafe {
         asm!("ud2", options(noreturn, nostack));
@@ -266,13 +266,14 @@ pub(super) const TLS_OFFSET: usize = 0;
 #[cfg(feature = "thread")]
 #[inline]
 pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usize) -> ! {
+    assert_eq!(__NR_exit, 60); // TODO: obviate this
     asm!(
         "syscall",
         "xor edi, edi",
-        "mov eax, {__NR_exit}",
+        "mov eax, 60", // TODO: use {__NR_exit}
         "syscall",
         "ud2",
-        __NR_exit = const __NR_exit,
+        //__NR_exit = const __NR_exit, // TODO: Use this when `asm_const` is stabilized.
         in("rax") __NR_munmap,
         in("rdi") map_addr,
         in("rsi") map_len,
@@ -280,23 +281,29 @@ pub(super) unsafe fn munmap_and_exit_thread(map_addr: *mut c_void, map_len: usiz
     );
 }
 
-/// Invoke the `__NR_rt_sigreturn` system call to return control from a signal
-/// handler.
-///
-/// # Safety
-///
-/// This function must never be called other than by the `sa_restorer`
-/// mechanism.
 #[cfg(feature = "signal")]
-#[naked]
-pub(super) unsafe extern "C" fn return_from_signal_handler() {
-    asm!(
-        "mov rax, {__NR_rt_sigreturn}",
-        "syscall",
-        "ud2",
-        __NR_rt_sigreturn = const __NR_rt_sigreturn,
-        options(noreturn)
-    );
+naked!(
+    "
+    Invoke the `__NR_rt_sigreturn` system call to return control from a signal
+    handler.
+
+    # Safety
+
+    This function must never be called other than by the `sa_restorer`
+    mechanism.
+    ";
+    pub(super) fn return_from_signal_handler() -> ();
+
+    "mov rax, 15", // TODO: use {__NR_rt_sigreturn}
+    "syscall",
+    "ud2";
+    //__NR_rt_sigreturn = const __NR_rt_sigreturn // TODO: Use this when `asm_const` is stabilized.
+    ;
+    options(noreturn)
+);
+#[cfg(test)] // TODO: obviate this
+fn test_rt_sigreturn() {
+    assert_eq!(__NR_rt_sigreturn, 15);
 }
 
 /// Invoke the appropriate system call to return control from a signal